backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic-ai==0.3.2
pydantic-ai-slim==0.3.2
pydantic-settings==2.5.2
openai==1.91.0